requirements.txt
pandas
numpy
scikit-learn
joblib
kafka-python
requests
python-dateutil
elasticsearch
pyyaml
